enabled to use custom Flexibility Env without using the gym format. deprecated --n_sample.
a custom FlexibilityEnv is created from the env name that's passed in. It has to follow the format FlexibilityNNxNNTt_SPx-v0 
if no SP is specified in the env, n_sample=5000 is used. In the first half of training, env with specified n_sample is used for training and testing. In the 2nd half of training, env with n_smaple=5000 is used for testing, while the env with specified n_sample is used for testing.